+++ src/core/utils/token-list.ts
@@ -1,6 +1,7 @@
+import {intersection} from '@exadel/esl/modules/esl-utils/misc/set';
+
 /** Class for processing attribute's tokens */
 export class TokenListUtils {
-
   /**
    * Divides string by whitespace regexp
    * @returns array of items or empty array
@@ -15,30 +16,20 @@ export class TokenListUtils {
   }
 
   /** Checks if all array elements are equal */
-  static hasSameElements(values: any[]): boolean {
+  static isAllEqual(values: any[]): boolean {
     return values.every((val) => val === values[0]);
-  }
-
-  /** Checks whether two arrays have same elements or not */
-  static equals<T>(arr1: T[], arr2: T[]): boolean {
-    return TokenListUtils.contains(arr1, arr2) && TokenListUtils.contains(arr2, arr1);
   }
 
   /** Checks if array contains all elements from subArray */
   static contains<T>(array: T[], subArray: T[]): boolean {
-    const set = new Set(array);
-    return subArray.every((val) => set.has(val));
-  }
-
-  /** Gets array which contains only common elements from arrays */
-  static intersection<T>(...rest: T[][]): T[];
-  static intersection<T>(a: T[], b: T[], ...rest: T[][]): T[] {
-    if (rest.length) return TokenListUtils.intersection(a, TokenListUtils.intersection(b, ...rest));
-    return b ? a.filter(Set.prototype.has, new Set(b)) : a;
+    return subArray.every((val) => array.includes(val));
   }
 
   /** Removes all element appearances from array */
   static remove<T>(array: T[], element: T): T[] {
     return array.filter((el) => el !== element);
   }
+
+  /** Returns intersection of all arrays */
+  static intersection = intersection;
 }