+++ src/plugins/editor/editor.tsx
@@ -8,30 +8,70 @@ import Prism from 'prismjs';
 import {CodeJar} from 'codejar';
 
 import {debounce} from '@exadel/esl/modules/esl-utils/async/debounce';
-import {bind, decorate, memoize} from '@exadel/esl/modules/esl-utils/decorators';
+import {boolAttr, decorate, listen, memoize} from '@exadel/esl/modules/esl-utils/decorators';
 
-import {UIPPlugin} from '../../core/base/plugin';
+import {UIPPluginPanel} from '../../core/panel/plugin-panel';
+import {CopyIcon} from '../copy/copy-button.icon';
+
+import {EditorIcon} from './editor.icon';
 
 /**
  * Editor {@link UIPPlugin} custom element definition
  * Uses Codejar code editor to provide an ability to modify UIP state markup
- * @extends UIPPlugin
  */
-export class UIPEditor extends UIPPlugin {
+export class UIPEditor extends UIPPluginPanel {
   public static override is = 'uip-editor';
+  public static override observedAttributes = ['copy', ...UIPPluginPanel.observedAttributes];
 
   /** Highlight method declaration  */
   public static highlight = (editor: HTMLElement): void => Prism.highlightElement(editor, false);
 
-  /** Wrapped {@link https://medv.io/codejar/ CodeJar} editor instance */
-  @memoize()
-  protected get editor(): CodeJar {
-    return CodeJar(this.$code, UIPEditor.highlight, { tab: '\t' });
+  /** Marker of JS Editor */
+  @boolAttr() public script: boolean;
+
+  /** Marker to display copy widget */
+  @boolAttr({name: 'copy'}) public showCopy: boolean;
+
+  protected override get $icon(): JSX.Element {
+    return <EditorIcon/>;
   }
 
   @memoize()
+  protected override get $toolbar(): HTMLElement {
+    const type = this.constructor as typeof UIPEditor;
+    return (
+      <div class={type.is + '-toolbar uip-plugin-header-toolbar'}>
+        {this.showCopy ? <uip-copy class={type.is + '-header-copy'}><CopyIcon/></uip-copy> : ''}
+      </div>
+    ) as HTMLElement;
+  }
+
+  /** {@link UIPEditor} section wrapper */
+  @memoize()
+  protected get $inner(): HTMLDivElement {
+    const type = this.constructor as typeof UIPPluginPanel;
+    return (
+      <div className={`${type.is}-inner uip-plugin-inner uip-plugin-inner-bg`}>
+        <esl-scrollbar class={type.is + '-scrollbar'} target="::next"/>
+        <div class={type.is + '-container esl-scrollable-content'}>
+          {this.$code}
+        </div>
+      </div>
+    ) as HTMLDivElement;
+  }
+
+  /** Code block */
+  @memoize()
   protected get $code(): HTMLElement {
-    return (<pre class='language-html editor-content'><code/></pre>) as HTMLElement;
+    const type = this.constructor as typeof UIPEditor;
+    const lang = this.script ? 'javascript' : 'html';
+    return (<pre class={type.is + '-code language-' + lang}><code/></pre>) as HTMLElement;
+  }
+
+  /** Wrapped [CodeJar](https://medv.io/codejar/) editor instance */
+  @memoize()
+  protected get editor(): CodeJar {
+    return CodeJar(this.$code, UIPEditor.highlight, {tab: '\t'});
   }
 
   /** @returns editor's content */
@@ -49,10 +89,9 @@ export class UIPEditor extends UIPPlugin {
     this.innerHTML = '';
 
     // Prefill content
+    this.appendChild(this.$header);
     this.appendChild(this.$inner);
-    this.$inner.classList.add('esl-scrollable-content');
-    this.$inner.append(<esl-scrollbar target="::parent"/>);
-    this.$inner.append(this.$code);
+    this.appendChild(this.$resize);
 
     // Initial update
     this._onRootStateChange();
@@ -66,16 +105,25 @@ export class UIPEditor extends UIPPlugin {
     super.disconnectedCallback();
   }
 
-  /** Callback to call on editor's content changes */
-  @decorate(debounce, 1000)
+  protected override attributeChangedCallback(attrName: string, oldVal: string, newVal: string): void {
+    super.attributeChangedCallback(attrName, oldVal, newVal);
+    this.$header.remove();
+    this.$toolbar.remove();
+    memoize.clear(this, ['$header', '$toolbar']);
+    this.insertAdjacentElement('afterbegin', this.$header);
+  }
+
+  /** Callback to call on an editor's content changes */
+  @decorate(debounce, 2000)
   protected _onChange(): void {
-    this.model!.setHtml(this.value, this);
+    if (this.script) this.model!.setJS(this.value, this);
+    else this.model!.setHtml(this.value, this);
   }
 
   /** Change editor's markup from markup state changes */
-  @bind
-  protected override _onRootStateChange(): void {
+  @listen({event: 'uip:change', target: ($this: UIPEditor) => $this.$root})
+  protected _onRootStateChange(): void {
     if (this.model!.lastModifier === this) return;
-    this.value = this.model!.html;
+    this.value = this.script ? this.model!.js : this.model!.html;
   }
 }