+++ src/plugins/registration.ts
@@ -1,18 +1,42 @@
-import {registeredSettings} from '../settings/registration';
+import {UIPSnippets} from './snippets/snippets';
+import {UIPSnippetsList} from './snippets-list/snippets-list';
+import {UIPSnippetsTitle} from './snippets-title/snippets-title';
+export {UIPSnippetsTitle, UIPSnippetsList, UIPSnippets};
 
-import {UIPHeader} from './header/header';
-import {UIPOptions} from './header/options/options';
-import {UIPOptionIcons} from './header/options/option-icons';
-import {UIPSnippets} from './header/snippets/snippets';
 import {UIPEditor} from './editor/editor';
-import {UIPSettings} from './settings/settings';
+export {UIPEditor};
 
-export {UIPOptions, UIPOptionIcons, UIPEditor, UIPSettings, UIPSnippets, UIPHeader};
+import {UIPSettings} from './settings/settings';
+import {UIPSetting} from './settings/base-setting/base-setting';
+import {UIPTextSetting} from './settings/text-setting/text-setting';
+import {UIPSelectSetting} from './settings/select-setting/select-setting';
+import {UIPBoolSetting} from './settings/bool-setting/bool-setting';
+import {UIPSliderSetting} from './settings/slider-setting/slider-setting';
+export {UIPSetting, UIPSettings, UIPTextSetting, UIPBoolSetting, UIPSelectSetting, UIPSliderSetting};
+
+import {UIPCopy} from './copy/copy-button';
+export {UIPCopy};
+
+import {UIPDirSwitcher} from './direction/dir-toggle';
+import {UIPThemeSwitcher} from './theme/theme-toggle';
+export {UIPDirSwitcher, UIPThemeSwitcher};
+
+export const registerSettings = (): void => {
+  UIPSettings.register();
+  UIPBoolSetting.register();
+  UIPTextSetting.register();
+  UIPSelectSetting.register();
+  UIPSliderSetting.register();
+};
 
 export const registerPlugins = (): void => {
-  UIPHeader.register();
-  UIPOptions.register();
+  UIPCopy.register();
+  UIPDirSwitcher.register();
+  UIPThemeSwitcher.register();
+
   UIPSnippets.register();
+  UIPSnippetsList.register();
+  UIPSnippetsTitle.register();
+
   UIPEditor.register();
-  registeredSettings().then(() => UIPSettings.register());
 };