+++ src/plugins/settings/settings.tsx
@@ -1,64 +1,113 @@
 import React from 'jsx-dom';
 
-import {bind, attr, listen} from '@exadel/esl/modules/esl-utils/decorators';
+import {isElement} from '@exadel/esl/modules/esl-utils/dom/api';
+import {debounce} from '@exadel/esl/modules/esl-utils/async/debounce';
+import {attr, boolAttr, decorate, listen, memoize} from '@exadel/esl/modules/esl-utils/decorators';
 
-import {UIPPlugin} from '../../core/base/plugin';
-// eslint-disable-next-line import/no-cycle
-import {UIPSetting} from './setting';
+import {UIPPluginPanel} from '../../core/panel/plugin-panel';
+import {ThemeToggleIcon} from '../theme/theme-toggle.icon';
+
+import {UIPSetting} from './base-setting/base-setting';
+import {SettingsIcon} from './settings.icon';
 
 /**
  * Settings {@link UIPPlugin} custom element definition
  * Container for {@link UIPSetting}
  */
-export class UIPSettings extends UIPPlugin {
+export class UIPSettings extends UIPPluginPanel {
   public static is = 'uip-settings';
+  public static observedAttributes = ['dir-toggle', 'theme-toggle', ...UIPPluginPanel.observedAttributes];
 
-  /**
-   * Attribute to set all inner {@link UIPSetting} settings'
-   * {@link UIPSetting#target} targets
-   */
+  /** Attribute to set all inner {@link UIPSetting} settings' {@link UIPSetting#target} targets */
   @attr() public target: string;
-  @attr({defaultValue: 'Settings'}) public label: string;
+
+  @boolAttr() public dirToggle: boolean;
+  @boolAttr() public themeToggle: boolean;
+
+  /** @readonly internal settings items state marker */
+  @boolAttr({readonly: true}) public inactive: boolean;
+
+  protected override get $icon(): JSX.Element {
+    return <SettingsIcon/>;
+  }
+
+  @memoize()
+  protected override get $toolbar(): HTMLElement {
+    const type = this.constructor as typeof UIPSettings;
+    return (<div class={type.is + '-toolbar uip-plugin-header-toolbar'}>
+      {this.themeToggle ? <uip-theme-toggle class={type.is + '-toolbar-option'}><ThemeToggleIcon/></uip-theme-toggle> : ''}
+      {this.dirToggle ? <uip-dir-toggle class={type.is + '-toolbar-option'}/> : ''}
+    </div>) as HTMLElement;
+  }
+
+  @memoize()
+  protected get $inner(): HTMLElement {
+    const type = this.constructor as typeof UIPSettings;
+    return (<div class={type.is + '-inner uip-plugin-inner uip-plugin-inner-bg'}>
+      <esl-scrollbar class={type.is + '-scrollbar'} target="::next"/>
+      {this.$container}
+    </div>) as HTMLElement;
+  }
+
+  @memoize()
+  protected get $container(): HTMLElement {
+    const type = this.constructor as typeof UIPSettings;
+    return (<div class={type.is + '-container esl-scrollable-content'}/>) as HTMLElement;
+  }
+
+  @memoize()
+  /** @returns HTMLElement[] - all internal items marked as settings item */
+  protected get $items(): HTMLElement[] {
+    return [...this.childNodes].filter(
+      ($el: ChildNode): $el is HTMLElement => isElement($el) && $el.hasAttribute('uip-settings-content')
+    );
+  }
+
+  /** @returns Element[] - active internal settings items */
+  protected get $activeItems(): Element[] {
+    return this.$items.filter(($el: Element) => !$el.classList.contains('uip-inactive-setting'));
+  }
 
   protected override connectedCallback(): void {
     super.connectedCallback();
-    this.updateInner();
-    this._onRootStateChange();
-  }
-
-  /** Initializes settings layout */
-  protected updateInner(): void {
-    const $content = <div className="settings-list esl-scrollable-content">
-      <esl-scrollbar target="::prev(.settings-list)"></esl-scrollbar>
-    </div>;
-    $content.append(...this.childNodes);
-    this.$inner.appendChild($content);
+    this.appendChild(this.$header);
     this.appendChild(this.$inner);
+    this.appendChild(this.$resize);
+    this.invalidate();
   }
 
-  protected disconnectedCallback(): void {
+  protected override disconnectedCallback(): void {
     super.disconnectedCallback();
+    this.append(...this.$items);
+    this.removeChild(this.$header);
+    this.removeChild(this.$inner);
   }
 
-  /**
-   * Handles `uip:change` event to
-   * apply changes to the state
-   */
-  @listen('uip:change')
-  protected _onSettingChanged(e: any): void {
-    e.stopPropagation();
-    if (!this.model) return;
-    (e.target as UIPSetting).applyTo(this.model);
+  protected override attributeChangedCallback(attrName: string, oldVal: string, newVal: string): void {
+    super.attributeChangedCallback(attrName, oldVal, newVal);
+    if (['label', 'collapsible', 'dir-toggle', 'theme-toggle'].includes(attrName)) {
+      this.$header.remove();
+      this.$toolbar.remove();
+      memoize.clear(this, ['$header', '$toolbar']);
+      this.insertAdjacentElement('afterbegin', this.$header);
+    }
   }
 
-  /** Collects all {@link UIPSetting} items */
-  protected get settings(): UIPSetting[] {
-    return Array.from(this.getElementsByClassName(UIPSetting.is)) as UIPSetting[];
+  @decorate(debounce, 100)
+  protected invalidate(): void {
+    const outside = this.$items.filter((el) => el.parentElement !== this.$container);
+    outside.forEach((el) => this.$container.appendChild(el));
   }
 
-  /** Updates {@link UIPSetting} values */
-  @bind
-  protected override _onRootStateChange(): void {
-    this.settings.forEach((setting) => setting.updateFrom(this.model!));
+  @listen('uip:settings:invalidate')
+  protected onInvalidate(): void {
+    this.invalidate();
+  }
+
+  /** Handles internal settings items state change */
+  @listen('uip:settings:state:change')
+  @decorate(debounce, 100)
+  protected onSettingsStateChange(): void {
+    this.$$attr('inactive', !this.$activeItems.length);
   }
 }