+++ src/registration.ts
@@ -1,10 +1,8 @@
 import {registerCore} from './core/registration';
-import {registerPlugins} from './plugins/registration';
-import {registerSettings} from './settings/registration';
+import {registerPlugins, registerSettings} from './plugins/registration';
 
 export * from './core/registration';
 export * from './plugins/registration';
-export * from './settings/registration';
 
 export function init(): void {
   registerCore();